requirements.txt
google-generativeai
PyYAML
python-dotenv
sqlalchemy
psycopg2-binary
supabase>=2.13.0
aiohttp>=3.8.0
pyyaml>=6.0.1
certifi>=2024.2.2